+++ include/robocup_nocom_pila/Describe_person_fmm.h
@@ -44,7 +44,6 @@ public:
     }
 */
 private:
-    
     int counter_;
     int count = 1;
 };
