+++ include/robocup_nocom_pila/Detect_person_cml.h
@@ -53,7 +53,7 @@ private:
     ros::Subscriber px_point_person;
 */
     float dist;
-    
+
     int counter_;
 };
