+++ include/robocup_nocom_pila/Go_home_cml.h
@@ -44,7 +44,6 @@ public:
     }*/
 
 private:
-
     int counter_;
 };
