+++ include/robocup_nocom_pila/Go_person_fmm.h
@@ -22,7 +22,7 @@
 #include "geometry_msgs/Twist.h"
 #include <sensor_msgs/Image.h>
 
-#include "../src/fsm_nav.cpp"
+#include "../src/nav.cpp"
 
 #include <string>
 
@@ -58,10 +58,9 @@ private:
     int person;
 
     int action = true;
+    int counter_;
     float dist;
     bool finish = false;
-    
-    int counter_;
 };
 
 }  // namespace robocup_nocom_pila
