+++ src/chatbot.cpp
@@ -1,3 +1,17 @@
+// Copyright 2022 Intelligent Robotics Lab
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #include <gb_dialog/DialogInterface.h>
 #include <string>
 #include "std_msgs/String.h"
@@ -36,8 +50,8 @@ class ExampleDF: public DialogInterface
   private:
     ros::NodeHandle nh_;
     ros::Publisher sound_pub;
-    
 };
+
 }  // namespace gb_dialog
 
 int main(int argc, char** argv)
