+++ src/robocup_nocom_pila/Detect_person_fmm.cpp
@@ -23,10 +23,10 @@ namespace robocup_nocom_pila
 {
 
 Detect_person_fmm::Detect_person_fmm(const std::string& name, const BT::NodeConfiguration & config)
-: BT::ActionNodeBase(name, config ), counter_(0)
+: BT::ActionNodeBase(name, config), counter_(0)
 {
   objects_bbx = nh.subscribe("/darknet_ros/bounding_boxes", 1, &Detect_person_fmm::DetectPersonBBXCallback, this);
-  objects_image = nh.subscribe("/camera/depth/image_raw", 1, &Detect_person_fmm::DetectPersonImageCallback, this);  
+  objects_image = nh.subscribe("/camera/depth/image_raw", 1, &Detect_person_fmm::DetectPersonImageCallback, this);
 }
 
 void
@@ -42,55 +42,51 @@ Detect_person_fmm::tick()
 
   sleep(0.5);
 
-  if(is_person == true && dist <= 2.5 && dist != 0)
+  if (is_person == true && dist <= 2.5 && dist != 0)
   {
     std::cerr << "HAY PERSONA" << std::endl;
     std::cerr << dist << std::endl;
     setOutput<int>("w_person", person);
+
     repeticiones = 0;
     person++;
     px = 0;
     py = 0;
     is_person = false;
     dist = 0;
-    return BT::NodeStatus::SUCCESS;
 
+    return BT::NodeStatus::SUCCESS;
   }
   else
   {
     std::cerr << "NO HAY PERSONA" << std::endl;
     std::cerr << dist << std::endl;
-    if(repeticiones >= 30)
+    if (repeticiones >= 30)
     {
       setOutput<int>("w_person", person);
       repeticiones = 0;
       person++;
-
     }
     std::cerr << repeticiones << std::endl;
-    
+
     repeticiones++;
 
     return BT::NodeStatus::RUNNING;
   }
-
-  
 }
 
 void Detect_person_fmm::DetectPersonBBXCallback(const darknet_ros_msgs::BoundingBoxesConstPtr& boxes)
 {
   for (const auto & box : boxes->bounding_boxes)
   {
-  
     if (box.Class == "person")
     {
       px = (box.xmax + box.xmin) / 2;
       py = (box.ymax + box.ymin) / 2;
       is_person = true;
-      //std::cerr << "yes, " << px << " , " << py << std::endl;
+      // std::cerr << "yes, " << px << " , " << py << std::endl;
     }
-    
-  } 
+  }
 }
 
 void Detect_person_fmm::DetectPersonImageCallback(const sensor_msgs::ImageConstPtr& image)
