+++ src/robocup_nocom_pila/PIDController.cpp
@@ -48,15 +48,21 @@ PIDController::get_output(double new_reference)
 
   // Proportional Error
   double direction = 0.0;
-  if (ref != 0.0) {
+  if (ref != 0.0)
+  {
     direction = ref / fabs(ref);
   }
 
-  if (fabs(ref) < min_ref_) {
+  if (fabs(ref) < min_ref_)
+  {
     output = 0.0;
-  } else if (fabs(ref) > max_ref_) {
+  }
+  else if (fabs(ref) > max_ref_)
+  {
     output = direction * max_output_;
-  } else {
+  }
+  else
+  {
     output = direction * min_output_ + ref * (max_output_ - min_output_);
   }
 
@@ -72,4 +78,4 @@ PIDController::get_output(double new_reference)
   return std::clamp(output, -max_output_, max_output_);
 }
 
-}  // namespace br2_tracking
+}  // namespace robocup_nocom_pila
