+++ src/robocup_nocom_pila/Speak_person_fmm.cpp
@@ -27,26 +27,23 @@ Speak_person_fmm::Speak_person_fmm(const std::string& name, const BT::NodeConfig
 : BT::ActionNodeBase(name, config), counter_(0)
 {
   // dist_sub = nh_.subscribe("/dist_person", 1, &Speak_person_fmm::PerceivePersonCallback, this);
-  //ear_sub = n_.subscribe("/sound/word", 1, &Speak_person_fmm::SpeakPersonCallback, this);
-
+  // ear_sub = n_.subscribe("/sound/word", 1, &Speak_person_fmm::SpeakPersonCallback, this);
 }
 
 void
 Speak_person_fmm::halt()
 {
-    ROS_INFO("Speak_person_fmm halt");
+  ROS_INFO("Speak_person_fmm halt");
 }
 
 BT::NodeStatus
 Speak_person_fmm::tick()
 {
-    ROS_INFO("Speak_person_fmm tick");
+  ROS_INFO("Speak_person_fmm tick");
 
-    forwarder.listen();
+  forwarder.listen();
 
-    
-
-    return BT::NodeStatus::RUNNING;
+  return BT::NodeStatus::RUNNING;
 }
 
 }  // namespace robocup_nocom_pila
